requirements.txt
azure-keyvault-secrets>=4.7.0
azure-identity>=1.14.0
snowflake-connector-python>=3.0.0
httpx>=0.24.0
pydantic>=2.0.0